+++ package.json
@@ -2,8 +2,10 @@
   "name": "todo-list",
   "jest": {
     "verbose": true,
-    "testEnvironment": "jsdom"
-},
+    "testEnvironment": "jsdom",
+    "resetMocks": false,
+    "setupFiles": ["./__setups__/localstorage.js", "jest-localstorage-mock"]
+  },
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
