+++ src/todo.js
@@ -1,5 +1,6 @@
 // eslint-disable-next-line
 import { updateToFalse, updateToTrue } from './update-status';
+import { deleteTodo } from './delete-todo';
 
 export const getAllTodos = () => {
   const todos = JSON.parse(localStorage.getItem('todos'));
@@ -60,18 +61,7 @@ export const renderTodo = () => {
   }
 
   deleteBtn.forEach((button) => {
-    button.addEventListener('click', (event) => {
-      const id = parseInt(event.target.getAttribute('data-index'), 10);
-      let localStoragetodos = getAllTodos();
-      localStoragetodos = localStoragetodos.filter((todo) => todo.index !== id);
-      // eslint-disable-next-line
-        for (let id = 0; id < localStoragetodos.length; id++) {
-        localStoragetodos[id].index = id;
-      }
-      localStorage.setItem('todos', JSON.stringify(localStoragetodos));
-      // eslint-disable-next-line
-        location.reload();
-    });
+    deleteTodo(button);
   });
 
   dropdownBtn.forEach((dropBtn) => {
