+++ src/todo.test.js
@@ -1,12 +1,16 @@
 import Todo from './todo-class';
 import { addTodo } from './todo';
+import { deleteTodo } from './delete-todo';
+import { getAllTodos } from './todo';
+import 'jest-localstorage-mock';
 
 
-  
-  describe('my beverage', () => {
+  describe('my todo task', () => {
 
     let todo = new Todo(false, "My task");
-    let result = addTodo(todo);
+    let todo1 = new Todo(false, "My task");
+    addTodo(todo);
+    addTodo(todo1);
 
     document.body.innerHTML =
     `
@@ -32,15 +36,17 @@ import { addTodo } from './todo';
        
         `
 
-    test('Add', () => {
-      expect(result[0].description).toBe("My task");
+    test('add', () => {
+    //  expect(result[0].description).toBe("My task");
+      expect(localStorage.getItem.mock.calls.length).toBe(2);
     });
 
     test('delete', () => {
         const deleteBtn = document.querySelectorAll('#delete');
         deleteBtn.forEach = (btn) => { 
-            btn.click();
+          deleteTodo(btn);
         }
+        expect(localStorage.getItem.mock.calls.length).toBe(0);
     })
 
   });
